+++ DearPyGui/src/Core/mvPythonTranslator.cpp
@@ -210,10 +210,10 @@ namespace Marvel {
 
 		PyObject* result = PyList_New(4);
 
-		PyList_SetItem(result, 0, ToPyFloat(color.r));
-		PyList_SetItem(result, 1, ToPyFloat(color.g));
-		PyList_SetItem(result, 2, ToPyFloat(color.b));
-		PyList_SetItem(result, 3, ToPyFloat(color.a));
+		PyList_SetItem(result, 0, ToPyFloat((float)color.r));
+		PyList_SetItem(result, 1, ToPyFloat((float)color.g));
+		PyList_SetItem(result, 2, ToPyFloat((float)color.b));
+		PyList_SetItem(result, 3, ToPyFloat((float)color.a));
 
 		return result;
 	}
